Return statements will now be generated if possible
If there is a Block, Do-loop or if-else-construct at the end of a method/function definition semantic check will now look recursively for a return statement. If non (or in case of if-else not enough) can be found a return statement will be placed at the end of the execution path if possible.
This is not possible for:
	- while loops (condition could be false right from start)
	- if-constructs (ditto)
	- print (simply don't know what to return)
	- break (that doesn't make sense either)
	- and of course everything that the parser rejected before
This is possible:
	- Binary ops (special case assignments: "x = a + b;" => "x = a + b; return x;")
	- unary ops
	- if-else, do, block (only if it is possible to place a return at the end recursively)